pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bulkRNASeq"
version = "0.1.0"
description = "A comprehensive bulk RNA-seq analysis pipeline"
readme = "README.md"
authors = [
    {name = "Mukul Sherekar", email = "example@example.com"}
]
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "snakemake>=7.32.4",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pylint>=3.0.0",
    "black>=23.9.1",
]

qc = [
    "fastqc>=0.12.1",
    "multiqc>=1.14",
]

align = [
    "hisat2>=2.2.1",
    "kallisto>=0.48.0",
]

counts = [
    "subread>=2.0.1",  # for featureCounts
]

post = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "biopython>=1.81",
    "gseapy>=1.0.0",
    "mygene>=3.2.2",
]

all = [
    "snakemake>=7.32.4",
    "pyyaml>=6.0.1",
    "fastqc>=0.12.1",
    "multiqc>=1.14",
    "hisat2>=2.2.1",
    "kallisto>=0.48.0",
    "subread>=2.0.1",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "biopython>=1.81",
    "gseapy>=1.0.0",
    "mygene>=3.2.2",
]

[project.scripts]
bulkrnaseq = "bulkRNASeq.main:main"



[project.urls]
Homepage = "https://github.com/yourusername/bulkRNASeq"
"Bug Tracker" = "https://github.com/yourusername/bulkRNASeq/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["bulkRNASeq*"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.pylint.messages_control]
disable = ["C0111", "C0103"]

[tool.conda-env]
name = "rnaseq"
channels = [
    "conda-forge",
    "bioconda",
    "defaults"
]
dependencies = [
    "fastqc=0.12.1",
    "cutadapt>=4.7",
    "subread=2.0.8",
    "samtools=1.21",
    "hisat2=2.2.1",
    "kallisto=0.51.1",
    "star",
    "multiqc",
    "htseq",
    "bioconductor-biocparallel=1.40.0",
    "bioconductor-deseq2",
    "snakemake",
    "scanpy=1.10.4",
    "anndata=0.11.3",
    "h5py=3.12.1",
    "xgboost=2.1.1",
    "r-base=4.4.2",
    "r-essentials=4.4",
    "r-tidyverse=2.0.0",
    "pytorch=2.5.1",
    "torchvision=0.20.1",
    "torchaudio=2.5.1"
]

[tool.conda-env.pip-dependencies]
packages = [
    "gseapy==0.9.5",
    "goatools",
    "umap-learn==0.5.7",
    "hdbscan==0.8.39",
    "legacy-api-wrap==1.4.1",
    "pynndescent==0.5.13"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 